Fix per file options by setting format in `on_load` hook
`ytdl-format` set as a per file option would make it impossible
to change the format (the menu would show the new format,
but it wouldn't actually change).
By setting it in the `on_load` hook before ytdl_hook runs, not does it
fix the problem with per file options, but we can use file-local-options
ourselves, making it unnecessary to keep track of the original format.

Another nice bonus is that now it's not necessary to reload upon
resetting the format, making video changes faster (in theory).