Move fetching formats on start into `file-loaded` event
Opening a playlist would call ytdl for the playlist itself, just to be
killed as soon as mpv opens the first file in the playlist.
`file-loaded` doesn't get called unless mpv can actually open the file,
which isn't the case for playlists.

This also shortens our `start-file` event, which apparently runs before
the `on_load` hook, making everything wait for it's completion.